% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Vibe-Based Playlist Generator App}
\usage{
launch_app()
}
\description{
This function launches the Shiny app for generating music playlists based on song "vibe" similarity.
}
