% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_function1.R
\name{random_playlist_length}
\alias{random_playlist_length}
\title{Generate a Random Playlist Length}
\usage{
random_playlist_length(mean = 20, sd = 5)
}
\arguments{
\item{mean}{The average number of songs you want (default = 20).}

\item{sd}{The standard deviation (default = 5).}
}
\value{
An integer for playlist length.
}
\description{
This function generates a random integer representing the number of songs in a playlist.
}
\examples{
random_playlist_length()
random_playlist_length(mean = 30)
random_playlist_length(mean = 25, sd = 10)
}
